+++ packages/next/src/index.ts
@@ -280,7 +280,6 @@ export const build: BuildV2 = async buildOptions => {
     cliType,
     lockfileVersion,
     packageJsonPackageManager,
-    nodeVersion,
     env: spawnOpts.env || {},
     turboSupportsCorepackHome,
     projectCreatedAt: config.projectSettings?.createdAt,
@@ -391,7 +390,6 @@ export const build: BuildV2 = async buildOptions => {
             [],
             spawnOpts,
             meta,
-            nodeVersion,
             config.projectSettings?.createdAt
           );
         }
@@ -1104,7 +1102,6 @@ export const build: BuildV2 = async buildOptions => {
       ['--production'],
       spawnOpts,
       meta,
-      nodeVersion,
       config.projectSettings?.createdAt
     );
   }