+++ packages/frameworks/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/frameworks
 
+## 3.14.0
+
+### Minor Changes
+
+- Add zero-config support for Fastify ([#14120](https://github.com/vercel/vercel/pull/14120))
+
+## 3.13.0
+
+### Minor Changes
+
+- Add root main as entrypoint for NestJS ([#14105](https://github.com/vercel/vercel/pull/14105))
+
 ## 3.12.0
 
 ### Minor Changes